requirements.txt
requests==2.21.0
python-dotenv==0.10.1
Flask==1.0.2
gunicorn==19.9.0
Flask-SQLAlchemy==2.3.2
Flask-MySQLdb==0.2.0
youtube-dl==2019.3.18
streamlink==1.1.1